# contracts/openapi.v1.yaml
openapi: 3.0.3
info:
  title: PM Agent API
  version: 1.0.0
  description: |
    Versioned contract between the Next.js frontend and FastAPI backend.
    Breaking changes require /v2. Authenticated routes send `Authorization: Bearer <access_token>`
    and optional `X-Workspace-Id`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /workspaces:
    get:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Workspace list } }
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses: { "201": { description: Created } }
  /team:
    get:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Members } }
  /team/invites:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Invited } }
  /team/{member_id}:
    delete:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: member_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Removed } }
  /onboarding:
    get:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Onboarding state } }
    patch:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Updated } }
  /onboarding/connect-source:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      summary: Verify source credentials, store them, and import recent tickets
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectSourceRequest" }
      responses: { "200": { description: Source connected } }
  /onboarding/sync-source:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      summary: Pull new tickets from a connected source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties: { provider: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceSyncResponse" }
  /onboarding/connect-repo:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Repo connected } }
  /onboarding/connect-output:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Output connected } }
  /onboarding/complete:
    post:
      tags: [workspaces]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Setup done } }
...
  /pipeline:
    get:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Kanban cards } }
  /pipeline/{card_id}:
    patch:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: card_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Moved } }
  /pipeline/{card_id}/advance:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: card_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Advanced } }
  /connections:
    get:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Sources, repos, outputs } }
  /connections/ingest:
    get:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Per-workspace ingest URL and secret } }
  /connections/sources:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Source added } }
  /connections/sources/{connection_id}:
    delete:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: connection_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Removed } }
  /connections/outputs:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Output added } }
  /connections/outputs/{connection_id}:
    delete:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: connection_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Removed } }
  /connections/repos:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Repo added } }
  /connections/repos/{repo_id}:
    delete:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: repo_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Removed } }
  /connections/repos/{repo_id}/reindex:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: repo_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Reindexed } }
  /connections/dev-agent:
    patch:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Toggled } }
  /product-docs:
    get:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Docs } }
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Confirmed } }
  /product-docs/presign:
    post:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Presigned PUT URL } }
  /product-docs/{doc_id}:
    delete:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: doc_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Deleted } }
...
  /webhooks/ingest/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: Ticket ingested } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ConnectSourceRequest:
      type: object
      required: [provider]
      properties:
        provider: { type: string }
        account_label: { type: string }
        domain: { type: string }
        api_key: { type: string }
        email: { type: string }
        password: { type: string }
        imap_host: { type: string }
        spreadsheet_id: { type: string }
        sheet_name: { type: string }
        instance_url: { type: string }
        client_id: { type: string }
        client_secret: { type: string }
        security_token: { type: string }
    SourceSyncResponse:
      type: object
      properties:
        provider: { type: string }
        imported: { type: integer }
        skipped: { type: integer }
...